+++ koin-projects/examples/androidx-samples/src/main/java/org/koin/sample/androidx/mvvm/MVVMActivity.kt
@@ -7,10 +7,10 @@ import org.junit.Assert.*
 import org.koin.android.ext.android.getKoin
 import org.koin.androidx.fragment.android.setupKoinFragmentFactory
 import org.koin.androidx.scope.lifecycleScope
+import org.koin.androidx.viewmodel.ViewModelOwner.Companion.from
+import org.koin.androidx.viewmodel.ViewModelOwner.Companion.fromAny
 import org.koin.androidx.viewmodel.ext.android.getViewModel
-import org.koin.androidx.viewmodel.ext.android.stateViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
-import org.koin.androidx.viewmodel.scope.stateViewModel
 import org.koin.androidx.viewmodel.scope.viewModel
 import org.koin.core.parameter.parametersOf
 import org.koin.core.qualifier.named
@@ -25,27 +25,23 @@ import org.koin.sample.androidx.utils.navigateTo
 
 class MVVMActivity : AppCompatActivity() {
 
-    val simpleViewModel: SimpleViewModel by viewModel(clazz = SimpleViewModel::class) {
-        parametersOf(
-            ID
-        )
-    }
+    val simpleViewModel: SimpleViewModel by viewModel(clazz = SimpleViewModel::class) { parametersOf(ID) }
 
     val vm1: SimpleViewModel by viewModel(named("vm1")) { parametersOf("vm1") }
     val vm2: SimpleViewModel by viewModel(named("vm2")) { parametersOf("vm2") }
 
-    val scopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(this)
-    val extScopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(this, named("ext"))
+    val scopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(owner = { from(this) })
+    val extScopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(owner = { from(this) },
+        qualifier = named("ext"))
 
-    val bundleStateVm = Bundle().apply { putString("vm1", "value to stateViewModel") }
-    val savedVm: SavedStateViewModel by stateViewModel(bundle = { bundleStateVm }) { parametersOf("vm1") }
+    val bundle = Bundle().apply { putString("vm1", "value to stateViewModel") }
+    val savedVm: SavedStateViewModel by viewModel(state = { bundle }) { parametersOf("vm1") }
 
-    val bundleStateScope = Bundle().apply {
-        putString("vm2", "value to lifecycleScope.stateViewModel")
-    }
-    val scopedSavedVm: SavedStateViewModel by lifecycleScope.stateViewModel(this, named("vm2"), bundle = { bundleStateScope }) {
-        parametersOf("vm2")
-    }
+    val bundleStateScope = Bundle().apply { putString("vm2", "value to lifecycleScope.stateViewModel") }
+    val scopedSavedVm: SavedStateViewModel by lifecycleScope.viewModel(owner = { fromAny(this) },
+        qualifier = named("vm3"),
+        state = { bundleStateScope }
+    ) { parametersOf("vm3") }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         // should set `lifecycleScope` here because we're
@@ -83,6 +79,5 @@ class MVVMActivity : AppCompatActivity() {
         mvvm_button.setOnClickListener {
             navigateTo<ScopedActivityA>(isRoot = true)
         }
-
     }
 }