+++ koin-projects/koin-androidx-viewmodel/src/main/java/org/koin/androidx/viewmodel/ViewModelResolver.kt
@@ -2,6 +2,8 @@ package org.koin.androidx.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import org.koin.androidx.viewmodel.factory.DefaultViewModelFactory
+import org.koin.androidx.viewmodel.factory.StateViewModelFactory
 import org.koin.core.qualifier.Qualifier
 import org.koin.core.scope.Scope
 
@@ -11,9 +13,9 @@ internal fun <T : ViewModel> ViewModelProvider.resolveInstance(viewModelParamete
 }
 
 internal fun <T : ViewModel> ViewModelProvider.get(
-        viewModelParameters: ViewModelParameter<T>,
-        qualifier: Qualifier?,
-        javaClass: Class<T>
+    viewModelParameters: ViewModelParameter<T>,
+    qualifier: Qualifier?,
+    javaClass: Class<T>
 ): T {
     return if (viewModelParameters.qualifier != null) {
         get(qualifier.toString(), javaClass)
@@ -23,14 +25,17 @@ internal fun <T : ViewModel> ViewModelProvider.get(
 }
 
 internal fun <T : ViewModel> Scope.createViewModelProvider(
-        viewModelParameters: ViewModelParameter<T>
+    viewModelParameters: ViewModelParameter<T>
 ): ViewModelProvider {
     return ViewModelProvider(
-            viewModelParameters.viewModelStore,
-            if (viewModelParameters.bundle != null) {
-                stateViewModelFactory(viewModelParameters)
-            } else {
-                defaultViewModelFactory(viewModelParameters)
-            }
+        viewModelParameters.viewModelStore, pickFactory(viewModelParameters)
     )
+}
+
+private fun <T : ViewModel> Scope.pickFactory(
+    viewModelParameters: ViewModelParameter<T>): ViewModelProvider.Factory {
+    return when {
+        viewModelParameters.registryOwner != null -> StateViewModelFactory(this, viewModelParameters)
+        else -> DefaultViewModelFactory(this, viewModelParameters)
+    }
 }