+++ koin-projects/koin-androidx-viewmodel/src/main/java/org/koin/androidx/viewmodel/koin/KoinExt.kt
@@ -1,8 +1,9 @@
 package org.koin.androidx.viewmodel.koin
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelStoreOwner
+import org.koin.androidx.viewmodel.ViewModelOwnerDefinition
 import org.koin.androidx.viewmodel.ViewModelParameter
+import org.koin.androidx.viewmodel.scope.BundleDefinition
 import org.koin.androidx.viewmodel.scope.getViewModel
 import org.koin.core.Koin
 import org.koin.core.parameter.ParametersDefinition
@@ -10,28 +11,31 @@ import org.koin.core.qualifier.Qualifier
 import kotlin.reflect.KClass
 
 inline fun <reified T : ViewModel> Koin.viewModel(
-    owner: ViewModelStoreOwner,
     qualifier: Qualifier? = null,
+    noinline state: BundleDefinition? = null,
+    noinline owner: ViewModelOwnerDefinition,
     noinline parameters: ParametersDefinition? = null
 ): Lazy<T> {
-    return lazy(LazyThreadSafetyMode.NONE) { getViewModel<T>(owner, qualifier, parameters) }
+    return lazy(LazyThreadSafetyMode.NONE) { getViewModel<T>(qualifier, state, owner, parameters) }
 }
 
 inline fun <reified T : ViewModel> Koin.getViewModel(
-    owner: ViewModelStoreOwner,
     qualifier: Qualifier? = null,
+    noinline state: BundleDefinition? = null,
+    noinline owner: ViewModelOwnerDefinition,
     noinline parameters: ParametersDefinition? = null
 ): T {
-    return getViewModel(owner, T::class, qualifier, parameters)
+    return getViewModel(qualifier, state, owner, T::class, parameters)
 }
 
 fun <T : ViewModel> Koin.getViewModel(
-    owner: ViewModelStoreOwner,
-    clazz: KClass<T>,
     qualifier: Qualifier? = null,
+    state: BundleDefinition? = null,
+    owner: ViewModelOwnerDefinition,
+    clazz: KClass<T>,
     parameters: ParametersDefinition? = null
 ): T {
-    return _scopeRegistry.rootScope.getViewModel(owner, clazz, qualifier, parameters)
+    return _scopeRegistry.rootScope.getViewModel(qualifier, state, owner, clazz, parameters)
 }
 
 fun <T : ViewModel> Koin.getViewModel(viewModelParameters: ViewModelParameter<T>): T {